Изменения Branch - Taalai
1) Поменял com.example на com.company
2) В таблице User добавил поле Gender
3) Добавил enum Gender (Male, Female)
4) Добавил зависимости JWT Token
5) Server-port 3001
6) Названия базы поменял на db_realestate
7) Не забудьте поменять пароль от PostgreSQL
