+++ .idea/compiler.xml
@@ -2,10 +2,6 @@
 <project version="4">
   <component name="CompilerConfiguration">
     <annotationProcessing>
-      <profile default="true" name="Default" enabled="true" />
-<<<<<<< HEAD
-    </annotationProcessing>
-=======
       <profile name="Maven default annotation processors profile" enabled="true">
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
@@ -21,6 +17,5 @@
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
       <module name="realestatesite" options="-parameters" />
     </option>
->>>>>>> origin/main
   </component>
 </project>
