+++ .idea/jpa-buddy.xml
@@ -1,10 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="JpaBuddyIdeaProjectConfig">
-<<<<<<< HEAD
-=======
-    <option name="defaultUnitInitialized" value="true" />
->>>>>>> origin/main
     <option name="renamerInitialized" value="true" />
   </component>
 </project>
