+++ server/src/main/java/com/company/realestatesite/controllers/UserController.java
@@ -1,0 +1,47 @@
+package com.company.realestatesite.controllers;
+
+import com.company.realestatesite.models.dtos.UserDto;
+import com.company.realestatesite.services.UserService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@Api(tags = "2. Пользователь")
+@RestController
+@RequestMapping(value = "/api/v1/user")
+@RequiredArgsConstructor
+public class UserController {
+
+    private final UserService userService;
+
+    @GetMapping("/find/by/id")
+    @ApiOperation("Поиск пользователя по id")
+    ResponseEntity<?> findById(@RequestParam Long id) {
+        try {
+            return new ResponseEntity<>(userService.findById(id), HttpStatus.FOUND);
+        } catch (Exception e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+        }
+    }
+    @GetMapping("/find/all")
+    @ApiOperation("Вывод всех пользователей")
+    ResponseEntity<List<UserDto>> findAll() {
+        return ResponseEntity.ok(userService.findAll());
+    }
+
+    @DeleteMapping("/delete")
+    @ApiOperation("Удаления пользователя")
+    ResponseEntity<?> delete(@RequestParam Long id) {
+        try {
+            userService.delete(id);
+            return ResponseEntity.ok("database entry deleted");
+        } catch (Exception e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+        }
+    }
+}
