+++ server/src/main/java/com/company/realestatesite/models/dtos/BaseDto.java
@@ -1,0 +1,24 @@
+package com.company.realestatesite.models.dtos;
+
+import com.company.realestatesite.models.enums.Status;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+import java.util.Date;
+
+@Getter
+@Setter
+@JsonIgnoreProperties(ignoreUnknown = true)
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class BaseDto {
+    private Long id;
+    @JsonFormat(pattern = "dd.MM.yyyy")
+    private Date created;
+    @JsonFormat(pattern = "dd.MM.yyyy")
+    private Date updated;
+    private Status status;
+}
