+++ server/src/main/java/com/company/realestatesite/models/entities/Role.java
@@ -1,0 +1,41 @@
+package com.company.realestatesite.models.entities;
+
+import com.company.realestatesite.models.enums.Status;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.sun.istack.NotNull;
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+import javax.persistence.*;
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@Entity
+@Table(name = "tb_role")
+public class Role {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "status")
+    private Status status;
+
+    @Column(name = "name", unique = true)
+    @NotNull
+    private String name;
+
+    @JsonIgnore
+    @ManyToMany(mappedBy = "roles", fetch = FetchType.LAZY)
+    private List<User> users;
+
+    @PrePersist
+    protected void onCreate() {
+        status = Status.ACTIVE;
+    }
+}
