+++ server/src/main/java/com/company/realestatesite/models/enums/Gender.java
@@ -1,0 +1,7 @@
+package com.company.realestatesite.models.enums;
+
+public enum Gender {
+
+    MALE,
+    FEMALE;
+}
