+++ server/src/main/java/com/company/realestatesite/models/enums/Status.java
@@ -1,0 +1,7 @@
+package com.company.realestatesite.models.enums;
+
+public enum Status {
+    ACTIVE,
+    NOT_ACTIVE,
+    DELETED
+}
