+++ server/src/main/java/com/company/realestatesite/realestatesiteApplication.java
@@ -1,0 +1,12 @@
+package com.company.realestatesite;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class realestatesiteApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(realestatesiteApplication.class, args);
+    }
+}
