+++ server/src/main/java/com/company/realestatesite/repository/AnnouncementRepository.java
@@ -1,0 +1,9 @@
+package com.company.realestatesite.repository;
+
+import com.company.realestatesite.models.entities.Announcement;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AnnouncementRepository extends JpaRepository<Announcement, Long> {
+}
