+++ server/src/main/java/com/company/realestatesite/repository/RoleRepository.java
@@ -1,0 +1,10 @@
+package com.company.realestatesite.repository;
+
+import com.company.realestatesite.models.entities.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Role findByName(String name);
+}
