+++ server/src/main/java/com/company/realestatesite/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.company.realestatesite.repository;
+
+import com.company.realestatesite.models.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    User findByUsername(String name);
+    User findByEmail(String email);
+}
