--- server/src/main/java/com/example/realestatesite/models/enums/Status.java
@@ -1,7 +1,0 @@
-package com.example.realestatesite.models.enums;
-
-public enum Status {
-    ACTIVE,
-    NOT_ACTIVE,
-    DELETED
-}
