--- server/src/main/java/com/example/realestatesite/realestatesiteApplication.java
@@ -1,12 +1,0 @@
-package com.example.realestatesite;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class realestatesiteApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(realestatesiteApplication.class, args);
-    }
-}
