--- server/src/main/java/com/example/realestatesite/repository/AnnouncementRepository.java
@@ -1,9 +1,0 @@
-package com.example.realestatesite.repository;
-
-import com.example.realestatesite.models.entities.Announcement;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AnnouncementRepository extends JpaRepository<Announcement, Long> {
-}
