--- server/src/main/java/com/example/realestatesite/repository/RoleRepository.java
@@ -1,10 +1,0 @@
-package com.example.realestatesite.repository;
-
-import com.example.realestatesite.models.entities.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Role findByName(String name);
-}
