--- server/src/main/java/com/example/realestatesite/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.example.realestatesite.repository;
-
-import com.example.realestatesite.models.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findByUsername(String name);
-    User findByEmail(String email);
-}
