+++ server/src/main/resources/application.properties
@@ -1,8 +1,7 @@
 # Datasource
-spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
+spring.datasource.url=jdbc:postgresql://localhost:5432/db_realestate
 spring.datasource.username=postgres
-spring.datasource.password=pass
-#spring.datasource.hikari.schema=habit_tracker
+spring.datasource.password=postgres
 spring.datasource.driverClassName=org.postgresql.Driver
 spring.datasource.hikari.auto-commit=true
 spring.mvc.pathmatch.matching-strategy=ant_path_matcher
@@ -19,3 +18,10 @@ logging.level.org.hibernate.type=TRACE
 
 # Log
 logging.level.org.springframework.boot.autoconfigure=INFO
+
+# Server-port
+server.port=3001
+
+# JWT
+jwt.token.secret=404E635266556A586E3272357538782F413F4428472B4B6250645367566B5970
+jwt.token.expired=86400000
